+++ src/main.rs
@@ -1,299 +1,21 @@
+mod reader;
+mod framestack;
+
 use std::cmp::max;
 use std::cmp::min;
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::fs::File;
-use std::io::BufRead;
 use std::io::BufReader;
 use std::rc::Rc;
 
-#[derive(Debug)]
-struct Tokens {
-    lines_buffer: Vec<BufReader<File>>,
-    token_buffer: Vec<String>,
-    imported_files: HashSet<String>,
-}
+use framestack::FrameStack;
 
-//since statement may be used multiple times when applying substitution
-// use Rc
-type Statement = Rc<[LanguageToken]>; //may be better to newtype this but I guess it works for now
+use crate::reader::Label;
+use crate::reader::Statement;
+use crate::framestack::Assertion;
+use crate::reader::{LanguageToken, Tokens, Proof};
 
-type Proof = Vec<Label>; //I don't think a proof is used multiple times
-type Label = Rc<str>;
-impl Tokens {
-    fn new(lines: BufReader<File>) -> Tokens {
-        Tokens {
-            lines_buffer: vec![lines],
-            token_buffer: vec![],
-            imported_files: HashSet::new(),
-        }
-    }
-    fn read(&mut self) -> Option<String> {
-        // println!("inside read function with state {:?}", self);
-        while self.token_buffer.is_empty() {
-            //println!("Buffer is empty, refilling");
-            let mut line = String::new();
-            // pretend this succeeeds
-            let result = self.lines_buffer.last_mut().unwrap().read_line(&mut line);
-            // println!("Read line: {}", line);
 
-            match result {
-                Ok(num) if num > 0 => {
-                    // println!("Read {} lines ", num);
-                    self.token_buffer = line.split_whitespace().map(|x| x.into()).collect();
-                    self.token_buffer.reverse();
-                }
-                _ => {
-                    // println!("Done with file");
-                    self.lines_buffer.pop();
-                    if self.lines_buffer.is_empty() {
-                        return None;
-                    }
-                }
-            }
-            // println!("Created token buffer {:?}", self.token_buffer);
-        }
-        self.token_buffer.pop()
-    }
-
-    fn read_file(&mut self) -> Option<String> {
-        // println!("reading file");
-
-        let mut token = self.read();
-        // println!("In read file found token {:?}", token);
-        while let Some("$[") = token.as_deref() {
-            let filename = self.read().expect("Couldn't find filename");
-            let endbracket = self.read().expect("Coludn't find end bracket");
-
-            // println!("In read file found filename: {:?}, endbracket: {:?}", filename, endbracket);
-            if endbracket != "$]" {
-                panic!("End bracket not found");
-            }
-
-            if !self.imported_files.contains(&filename) {
-                // println!("Found new file {}", &filename);
-
-                self.lines_buffer.push(BufReader::new(
-                    File::open(filename.clone()).expect("Failed to open file"),
-                ));
-                self.imported_files.insert(filename);
-            }
-            token = self.read();
-        }
-        token
-    }
-
-    fn read_comment(&mut self) -> Option<String> {
-        // println!("reading comment");
-
-        loop {
-            let mut token = self.read_file();
-            // println!("In read comment: found token to be {:?}", token);
-            match &token {
-                None => return None,
-                Some(x) if x == "$(" => loop {
-                    match token.as_deref() {
-                        Some("$)") => break,
-                        _ => token = self.read(),
-                    }
-                },
-                _ => return token,
-            }
-        }
-    }
-
-    fn readstat(&mut self) -> Statement {
-        let mut stat: Vec<Rc<str>> = vec![];
-        let mut token = self
-            .read_comment()
-            .expect("Failed to read token in read stat");
-
-        // println!("In read stat, found token to be {:?}", token);
-        while token != "$." {
-            stat.push(token.into());
-            token = self.read_comment().expect("EOF before $.");
-        }
-        stat.into()
-    }
-}
-
-type LanguageToken = Rc<str>;
-
-#[derive(Default, Debug)]
-struct Frame {
-    c: HashSet<LanguageToken>,
-    v: HashSet<LanguageToken>,
-    d: HashSet<(LanguageToken, LanguageToken)>, //maybe switch this give c and v different types
-    f: Vec<(LanguageToken, LanguageToken)>,
-    f_labels: HashMap<LanguageToken, Label>,
-    e: Vec<Statement>,
-    e_labels: HashMap<Statement, Label>,
-}
-
-#[derive(Default, Debug)]
-struct FrameStack {
-    list: Vec<Frame>,
-}
-
-#[derive(Debug)]
-struct Assertion {
-    dvs: HashSet<(LanguageToken, LanguageToken)>,
-    f_hyps: VecDeque<(LanguageToken, LanguageToken)>,
-    e_hyps: Vec<Statement>,
-    stat: Statement,
-}
-
-impl FrameStack {
-    fn push(&mut self) {
-        self.list.push(Frame::default());
-    }
-
-    fn add_c(&mut self, token: LanguageToken) {
-        let frame = &mut self.list.last_mut().unwrap();
-
-        if frame.c.contains(&token) {
-            panic!("Const already defined")
-        }
-        if frame.v.contains(&token) {
-            panic!("consta elaryd defined as var in scope")
-        }
-        frame.c.insert(token);
-    }
-
-    fn add_v(&mut self, token: LanguageToken) {
-        let frame = &mut self.list.last_mut().unwrap();
-
-        if frame.c.contains(&token) {
-            panic!("Variable already defined")
-        }
-        if frame.v.contains(&token) {
-            panic!("Variable elaryd defined as var in scope")
-        }
-        frame.v.insert(token);
-    }
-
-    fn add_f(&mut self, var: LanguageToken, kind: LanguageToken, label: Label) {
-        if !self.lookup_v(&var) {
-            panic!("var not defined")
-        }
-        if !self.lookup_c(&kind) {
-            panic!("const not defined")
-        }
-
-        let frame = self.list.last_mut().unwrap();
-        if frame.f_labels.contains_key(&var) {
-            panic!("f already defined in scope")
-        }
-        frame.f.push((var.clone(), kind));
-        frame.f_labels.insert(var, label);
-    }
-
-    fn add_e(&mut self, stat: Statement, label: Label) {
-        let frame = self.list.last_mut().unwrap();
-
-        frame.e.push(stat.clone());
-        frame.e_labels.insert(stat, label);
-    }
-
-    fn add_d(&mut self, stat: Statement) {
-        let frame = self.list.last_mut().unwrap();
-        //let mut product_vec = vec!();
-        for x in stat.iter() {
-            for y in stat.iter() {
-                if x != y {
-                    frame
-                        .d
-                        .insert((min(x.clone(), y.clone()), max(x.clone(), y.clone())));
-                }
-            }
-        }
-    }
-
-    fn lookup_c(&self, token: &str) -> bool {
-        self.list.iter().rev().any(|fr| fr.c.contains(token))
-    }
-
-    fn lookup_v(&self, token: &str) -> bool {
-        self.list.iter().rev().any(|fr| fr.v.contains(token))
-    }
-
-    fn lookup_f(&self, var: LanguageToken) -> Label {
-        // println!("lookup {}", var);
-        let f = self
-            .list
-            .iter()
-            .rev()
-            .find(|frame| frame.f_labels.contains_key(&var))
-            .unwrap();
-
-        f.f_labels[&var].clone()
-    }
-
-    fn lookup_d(&mut self, x: LanguageToken, y: LanguageToken) -> bool {
-        self.list.iter().rev().any(|fr| {
-            fr.d.contains(&(min(x.clone(), y.clone()), max(x.clone(), y.clone())))
-        })
-    }
-
-    fn lookup_e(&self, stmt: Statement) -> Label {
-        let f = self
-            .list
-            .iter()
-            .rev()
-            .find(|frame| frame.e_labels.contains_key(&stmt))
-            .expect("Bad e");
-
-        f.e_labels[&stmt].clone()
-    }
-
-    fn make_assertion(&mut self, stat: Statement) -> Assertion {
-        let _frame = self.list.last_mut().unwrap();
-
-        let e_hyps: Vec<Statement> = self.list.iter().flat_map(|fr| fr.e.clone()).collect();
-
-        let chained = e_hyps.iter().chain(std::iter::once(&stat));
-
-        let mut mand_vars: HashSet<LanguageToken> = chained
-            .flat_map(|x| x.iter())
-            .filter(|tok| self.lookup_v(tok))
-            .cloned()
-            .collect(); //cloned should do a shallow copy
-
-        // println!("ma: \n mand_vars: {:?}, ", mand_vars);
-
-        let mut cartesian: HashSet<(LanguageToken, LanguageToken)> = HashSet::new();
-
-        for x in mand_vars.iter() {
-            for y in mand_vars.iter() {
-                cartesian.insert((x.clone(), y.clone()));
-            }
-        }
-
-        let dvs: HashSet<(LanguageToken, LanguageToken)> = self
-            .list
-            .iter()
-            .flat_map(|fr| fr.d.intersection(&cartesian))
-            .cloned()
-            .collect();
-
-        let mut f_hyps = VecDeque::new();
-        self.list.iter().rev().for_each(|fr| {
-            fr.f.iter().rev().for_each(|(v, k)| {
-                if mand_vars.contains(v) {
-                    f_hyps.push_front((k.clone(), v.clone()));
-                    mand_vars.remove(v);
-                }
-            });
-        });
-        // println!("ma: \n dvs: {:?}, f: {:?}, e_hyps: {:?}, stat: {:?}", dvs, f_hyps, e_hyps, stat);
-
-        Assertion {
-            dvs,
-            f_hyps,
-            e_hyps,
-            stat,
-        }
-    }
-}
 
 // first one is label type,
